tests: move util functions to extended utils
