+++ test/integration/resize_test.go
@@ -44,7 +44,7 @@ var _ = Describe("vary VM parameters: memory cpus, disk", Serial, Ordered, Label
 
 		It("login to cluster using crc-admin context", func() {
 
-			err := AddOCToPath()
+			err := util.AddOCToPath()
 			Expect(err).NotTo(HaveOccurred())
 
 			cmd := exec.Command("oc", "config", "use-context", "crc-admin")
@@ -62,20 +62,20 @@ var _ = Describe("vary VM parameters: memory cpus, disk", Serial, Ordered, Label
 		})
 
 		It("check VM's memory size", func() {
-			out, err := SendCommandToVM("cat /proc/meminfo")
+			out, err := util.SendCommandToVM("cat /proc/meminfo")
 			Expect(err).NotTo(HaveOccurred())
 			Expect(out).Should(MatchRegexp(`MemTotal:[\s]*11\d{6}`))
 		})
 
 		It("check VM's number of cpus", func() {
-			out, err := SendCommandToVM("cat /proc/cpuinfo")
+			out, err := util.SendCommandToVM("cat /proc/cpuinfo")
 			Expect(err).NotTo(HaveOccurred())
 			Expect(out).Should(MatchRegexp(`processor[\s]*\:[\s]*4`))
 			Expect(out).ShouldNot(MatchRegexp(`processor[\s]*\:[\s]*5`))
 		})
 
 		It("check VM's disk size", func() {
-			out, err := SendCommandToVM("df -h | grep sysroot")
+			out, err := util.SendCommandToVM("df -h | grep sysroot")
 			Expect(err).NotTo(HaveOccurred())
 			Expect(out).Should(MatchRegexp(`.*40G[\s].*[\s]/sysroot`))
 		})
@@ -93,20 +93,20 @@ var _ = Describe("vary VM parameters: memory cpus, disk", Serial, Ordered, Label
 		})
 
 		It("check VM's memory size", func() {
-			out, err := SendCommandToVM("cat /proc/meminfo")
+			out, err := util.SendCommandToVM("cat /proc/meminfo")
 			Expect(err).NotTo(HaveOccurred())
 			Expect(out).Should(MatchRegexp(`MemTotal:[\s]*12\d{6}`))
 		})
 
 		It("check VM's number of cpus", func() {
-			out, err := SendCommandToVM("cat /proc/cpuinfo")
+			out, err := util.SendCommandToVM("cat /proc/cpuinfo")
 			Expect(err).NotTo(HaveOccurred())
 			Expect(out).Should(MatchRegexp(`processor[\s]*\:[\s]*5`))
 			Expect(out).ShouldNot(MatchRegexp(`processor[\s]*\:[\s]*6`))
 		})
 
 		It("check VM's disk size", func() {
-			out, err := SendCommandToVM("df -h | grep sysroot")
+			out, err := util.SendCommandToVM("df -h | grep sysroot")
 			Expect(err).NotTo(HaveOccurred())
 			Expect(out).Should(MatchRegexp(`.*50G[\s].*[\s]/sysroot`))
 		})
@@ -143,13 +143,13 @@ var _ = Describe("vary VM parameters: memory cpus, disk", Serial, Ordered, Label
 		})
 
 		It("check VM's memory size", func() {
-			out, err := SendCommandToVM("cat /proc/meminfo")
+			out, err := util.SendCommandToVM("cat /proc/meminfo")
 			Expect(err).NotTo(HaveOccurred())
 			Expect(out).Should(MatchRegexp(`MemTotal:[\s]*9\d{6}`)) // there should be a check if cluster needs >9216MiB; it isn't there and mem gets scaled down regardless
 		})
 
 		It("check VM's number of cpus", func() {
-			out, err := SendCommandToVM("cat /proc/cpuinfo")
+			out, err := util.SendCommandToVM("cat /proc/cpuinfo")
 			Expect(err).NotTo(HaveOccurred())
 			Expect(out).Should(MatchRegexp(`processor[\s]*\:[\s]*3`))
 			Expect(out).ShouldNot(MatchRegexp(`processor[\s]*\:[\s]*4`))
@@ -157,7 +157,7 @@ var _ = Describe("vary VM parameters: memory cpus, disk", Serial, Ordered, Label
 
 		if runtime.GOOS != "darwin" {
 			It("check VM's disk size", func() {
-				out, err := SendCommandToVM("df -h | grep sysroot")
+				out, err := util.SendCommandToVM("df -h | grep sysroot")
 				Expect(err).NotTo(HaveOccurred())
 				Expect(out).Should(MatchRegexp(`.*50G[\s].*[\s]/sysroot`))
 			})
