+++ test/integration/utilities_test.go
@@ -5,15 +5,9 @@ import (
 	"io"
 	"os"
 	"os/exec"
-	"path/filepath"
-	"runtime"
 	"strings"
 	"time"
 
-	crcConfig "github.com/crc-org/crc/v2/pkg/crc/config"
-	"github.com/crc-org/crc/v2/pkg/crc/constants"
-	"github.com/crc-org/crc/v2/pkg/crc/machine"
-	"github.com/crc-org/crc/v2/pkg/crc/ssh"
 	"github.com/crc-org/crc/v2/test/extended/crc/cmd"
 	. "github.com/onsi/gomega"
 )
@@ -94,42 +88,4 @@ func RunCRCExpectFail(args ...string) (string, error) {
 	}
 
 	return stderr, nil
-}
-
-// Send command to CRC VM via SSH
-func SendCommandToVM(cmd string) (string, error) {
-	client := machine.NewClient(constants.DefaultName, false,
-		crcConfig.New(crcConfig.NewEmptyInMemoryStorage(), crcConfig.NewEmptyInMemorySecretStorage()),
-	)
-	connectionDetails, err := client.ConnectionDetails()
-	if err != nil {
-		return "", err
-	}
-	ssh, err := ssh.NewClient(connectionDetails.SSHUsername, connectionDetails.IP, connectionDetails.SSHPort, connectionDetails.SSHKeys...)
-	if err != nil {
-		return "", err
-	}
-	out, _, err := ssh.Run(cmd)
-	if err != nil {
-		return "", err
-	}
-	return string(out), nil
-}
-
-func AddOCToPath() error {
-	path := os.ExpandEnv("${HOME}/.crc/bin/oc:$PATH")
-	if runtime.GOOS == "windows" {
-		userHomeDir, err := os.UserHomeDir()
-		if err != nil {
-			return err
-		}
-		unexpandedPath := filepath.Join(userHomeDir, ".crc/bin/oc;${PATH}")
-		path = os.ExpandEnv(unexpandedPath)
-	}
-	err := os.Setenv("PATH", path)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
